Add flex container for book cards
Here I'm hoping to build a flex container that will hold all of a user's book cards. My goal is that the container holds ups to 3 cards per row then wraps to the next row and so on. It seems to work OK right now, but could use some refinement (smaller/bigger margins, different gaps, etc.). I'll hold off on this until later since I'm not completely finished with the card design.