Complete displayBookCards func
The first iteration of this function used hard-coded values for a book's info and only created a single book. I took that and made it dynamic by pulling a book's info from a book object, then made the function run once for each book object is a library's book array.