(Mostly) Finish New Book Button process
Ok, so I've got the new book button process mostly working. It doesn't have any validation at this point and for some reason it sends true with the readStatus checkbox no matter if it's checked or not? I'm not sure why... Anyway, this was pretty tough because I wasn't sure how to 1. grab form values to use in a JS function and 2. how to stop the form from submitting http verbs and redirecting the session. After a couple hours of research I learned how to stop the form submit using e.preventDefault();