Complete remove book button
The changes here are for step 5 of the project: "Add a button on each book’s display to remove the book from the library." This was an interesting piece because I had to figure out how to dynamically add both the button and the button's on click event each time a card was created. I also had to figure out a way to keep track of which card corresponds to which book in the array (ended up using a data-* attribute). One thing I got stuck on for a bit was that I tried to add the button's click event before I added the button to the DOM, which probably would have been okay if I hadn't been using .parentElement to access the parent card (for the data attribute). Because the button hadn't been added to the DOM yet, .parentElement kept returning null.