Finish change read status button
This is for step 6 of the project: "Add a button on each book’s display to change its read status." I used a similar approach that I used for step 5, the remove book button. I dynamically am creating the button, then adding it to the DOM, then adding the click event each time a card is created. I decided to move these 3 steps for both buttons into their own functions (setUp...) to keep the displayBookCard function short. One piece I got stuck on was getting the correct readStatus from the read? checkbox on the new book form. For some reason, no matter if the box was checked or not, the element would return the same value. So, instead of basing the status off of the element's value, I instead used .checked to see if it was checked or not.

At this point, all of the functionality is complete for this project (huzzah!). My next steps are to clean up the display. The book cards are not very pretty at this time (eg the change read status text doesn't even fit into the button lol....). Also my form doesn't have any validation set up. Now I should probably set that up, especially if this were to be a portfolio project. But in the interest of time, and since validation isn't requested in the project instructions, I'm going to skip it. If I decide to use this project in my portfolio, I'll come back at a later time and polish it up.