Finish card redesign
I changed the card so now it looks much better and is responsive to different length information. I have a minheight on each card so it will scale vertically as there is more information but a hard width so it will not change horizontally. I based the initial card dimensions (assuming the minheight isn't exceeded) off of a standard business card, 3.5 x 2.

My next step is to implement these design changes in the javascript code that is making the cards.